+++ src/cymongoose/__init__.py
@@ -50,7 +50,7 @@ from ._mongoose import (  # noqa: F401
 )
 from .aio import AsyncManager
 
-__version__ = "0.1.14"
+__version__ = "0.2.0"
 
 __all__ = [
     # Classes